requirements.txt
flask==2.3.3
flask-cors==4.0.0
werkzeug==2.3.7
gunicorn==21.2.0
# Dependencies for the new utilities
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
requests==2.31.0
